feat: allow parent to control themes
The current implementation sets the theme at a fixed value when getStorybookUI() is called. It seems
like an assumption that a project is expected to set the theme once for an app's entire run. For the
use case where the app's end user is allowed to change the theme at will, the current implementation
is simply not flexible enough.